Expand/collapse for element vectors
